+++ test-api/schema-paginated/Node.js
@@ -2,7 +2,8 @@ import { nodeDefinitions, fromGlobalId } from 'graphql-relay'
 
 import joinMonster from '../../src/index'
 const options = {
-  minify: process.env.MINIFY == 1
+  minify: process.env.MINIFY == 1,
+  aliasPrefix: process.env.ALIAS_PREFIX
 }
 const { PAGINATE } = process.env
 if (knex.client.config.client === 'mysql') {